back/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-shortcuts==1.6
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.5
PyJWT==1.7.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
